+++ CHANGELOG.rst
@@ -7,7 +7,11 @@ v3.6.0
 *Release date: In development*
 
 - Removed xcuitest deprecated get_window_size() method and replaced it with get_window_rect() in all mobile actions
+- Add text comparison methods based on AI libraries. To use them, install the `ai` extra dependency:
 
+.. code:: console
+
+    $ pip install toolium[ai]
 
 v3.5.0
 ------